+++ snowpark-checkpoints-hypothesis/test/integ/telemetry_compare_utils.py
